+++ services/users-services.js
@@ -1,15 +1,18 @@
-const uri = "mongodb://admin:secret@127.0.0.1:27017";
+const {User} = require("../models/User");
+const {MongoClient, ObjectId} = require('mongodb');
 
+const URI = "mongodb://admin:secret@127.0.0.1:27017";
+const DB_NAME = "petstore";
+const COLLECTION = "users";
 
-class UsersServices {
+class UserServices {
   static async get() {
 
-    const client = new MongoClient(uri);
-
+    const client = new MongoClient(URI);
     try {
       await client.connect();
-      const database = client.db("petstore");
-      const collection = database.collection("users");
+      const database = client.db(DB_NAME);
+      const collection = database.collection(COLLECTION);
 
       const allRows = await collection.find().toArray();
 
@@ -20,27 +23,33 @@ class UsersServices {
   }
 
   static async getById(id) {
-      const client = new MongoClient(uri);
-      try {
-        await client.connect();
-        const database = client.db("petstore");
-        const collection = database.collection("users");
-
-        return await collection.findOne({_id: new ObjectId(id)});
-       
-      } finally {
-        await client.close();
+    const client = new MongoClient(URI);
+    try {
+      if (!ObjectId.isValid(id)) {
+        throw new Error("Non valid ID");
       }
+
+      await client.connect();
+      const database = client.db(DB_NAME);
+      const collection = database.collection(COLLECTION);
+
+      return await collection.findOne({_id: new ObjectId(id)});
+
+    } finally {
+      await client.close();
     }
+  }
 
   static async post(email, password) {
-    const client = new MongoClient(uri);
+    const client = new MongoClient(URI);
     try {
       await client.connect();
-        const database = client.db("petstore");
-        const collection = database.collection("users");
+      const database = client.db(DB_NAME);
+      const collection = database.collection(COLLECTION);
 
-      return await collection.insertOne({email: email, password: password});
+      const item = new User(email, password);
+
+      return await collection.insertOne(item);
 
     } finally {
       await client.close();
@@ -50,11 +59,15 @@ class UsersServices {
 
 
   static async delete(id) {
-    const client = new MongoClient(uri);
+    const client = new MongoClient(URI);
     try {
-            await client.connect();
-        const database = client.db("petstore");
-        const collection = database.collection("users");
+      if (!ObjectId.isValid(id)) {
+        throw new Error("Non valid ID");
+      }
+      await client.connect();
+      const database = client.db(DB_NAME);
+      const collection = database.collection(COLLECTION);
+
 
       return await collection.deleteOne({_id: new ObjectId(id)});
 
@@ -64,11 +77,17 @@ class UsersServices {
   }
 
   static async update(email, password) {
-    const client = new MongoClient(uri);
+    const client = new MongoClient(URI);
     try {
-              await client.connect();
-        const database = client.db("petstore");
-        const collection = database.collection("users");
+
+      if (!ObjectId.isValid(id)) {
+        throw new Error("ID no válido");
+      }
+      await client.connect();
+      const database = client.db(DB_NAME);
+      const collection = database.collection(COLLECTION);
+
+      new User(email, password);
 
       return await collection.updateOne({_id: new ObjectId(id)}, {$set: {email: email, password: password}});
 
@@ -78,4 +97,4 @@ class UsersServices {
   }
 }
 
-module.exports = UsersServices;
+module.exports = UserServices;
